Moved the BrandingAppearance from the responses to the entities folder
