+++ src/ui/layout/navbar.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-  import { type BrandingAppearance } from "../../networking/responses/branding-response";
   import { Theme } from "../theme/theme";
   import NavBarHeader from "../navbar-header.svelte";
   import { onMount } from "svelte";
   import SectionLayout from "./section-layout.svelte";
+  import type { BrandingAppearance } from "../../entities/branding";
 
   export let brandingAppearance: BrandingAppearance | undefined = undefined;
   let style = new Theme(brandingAppearance).productInfoStyleVars;
